requirements.txt
PyYAML
tweepy
python-dotenv
